and fix merge conflicts

Conflicts:
	src/components/quest/examples/CMakeLists.txt
	src/components/sidre/src/interface/c_fortran/wrapfsidre.f

+++ src/components/sidre/src/core/DataView.hpp
@@ -26,6 +26,7 @@
 #include <set>
 
 // Other CS Toolkit headers
+#include "common/config.hpp"
 #include "common/CommonTypes.hpp"
 #include "slic/slic.hpp"
 
@@ -257,7 +258,7 @@ public:
    */
   int getNumDimensions() const
   {
-    return m_shape.size();
+    return static_cast<int>(m_shape.size());
   }
 
   /*!
